Reorganized Net classes
